fix(render): rename style pipe scoll.ts to scroll.ts
Correct the scroll style pipe filename and update manifest pipe references.
